Startup text when magnifier is off, app icon.
Add a large text message in the main magnifier window if the magnifier is off at startup. This is to help with the confusion I had thinking the magnifier was not working.
Add an icon for the app and use it on the forms also (control panel does not show the icon on the screen, but it's still there).